Updated iOS SDK to v5.5.5
