--- ios/Crittercism.framework/Crittercism
@@ -1,1 +1,0 @@
-Versions/Current/Crittercism
