--- ios/Crittercism.framework/Modules/module.modulemap
@@ -1,9 +1,0 @@
-framework module Crittercism {
-  umbrella header "Crittercism.h"
-  
-  export *
-  module * { export * }
-  
-  link framework "Foundation"
-  link framework "SystemConfiguration"
-}
