--- ios/Crittercism.framework/Versions/5_4_11/Headers/CrittercismLoggingLevel.h
@@ -1,14 +1,0 @@
-//
-//  CRLoggingLevel.h
-//  Crittercism-iOS
-//
-//  Created by David Shirley 2 on 8/12/15.
-//  Copyright © 2015 Crittercism. All rights reserved.
-//
-
-typedef NS_ENUM(NSInteger, CRLoggingLevel) {
-  CRLoggingLevelSilent = 0,
-  CRLoggingLevelError = 1,
-  CRLoggingLevelWarning = 2,
-  CRLoggingLevelInfo = 3,
-};
