--- ios/Crittercism.framework/Versions/Current
@@ -1,1 +1,0 @@
-5_4_11
