+++ ios/Crittercism/ExampleApp.html
@@ -1,0 +1,8 @@
+<html>
+   <head>
+      <meta http-equiv="refresh" content="0; url=https://github.com/crittercism/crittercism-example-apps">
+   </head>
+   <body>
+   <p>Click <a href="https://github.com/crittercism/crittercism-example-apps">here</a> if you are not redirected automatically.</p>
+   </body>
+</html>
