+++ ios/Crittercism/ReleaseNotes.html
@@ -1,0 +1,8 @@
+<html>
+   <head>
+      <meta http-equiv="refresh" content="0; url=http://docs.crittercism.com/release_notes/release_notes_ios.html">
+   </head>
+   <body>
+   <p>Click <a href="http://docs.crittercism.com/release_notes/release_notes_ios.html">here</a> if you are not redirected automatically.</p>
+   </body>
+</html>
