+++ ios/Crittercism/tvOS/Crittercism.framework/Crittercism
@@ -1,0 +1,1 @@
+Versions/Current/Crittercism
