+++ ios/CrittercismExampleApp/main.m
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-
-#import <UIKit/UIKit.h>
 #import "AppDelegate.h"
 
-int main(int argc, char * argv[]) {
+int main(int argc, char *argv[]) {
     @autoreleasepool {
         return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
     }
